Refactor DynamicList and ViewState components, introducing a new structure for dynamic lists with enhanced support for customizable row, detail, error, and skeleton views. Implement a builder pattern for easier configuration of dynamic lists and sectioned lists. Update documentation to reflect new features and usage examples, ensuring clarity for developers. Revise tests to validate the new functionality and integration of dynamic list components.
