+++ Sources/DynamicList/Core Components/Default Views/DefaultSkeletonView.swift
@@ -1,0 +1,34 @@
+//
+//  Copyright © 2025 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import SwiftUI
+
+/// Default skeleton view for loading states
+struct DefaultSkeletonView: View {
+    var body: some View {
+        List(0 ..< 10, id: \.self) { _ in
+            HStack {
+                RoundedRectangle(cornerRadius: 4)
+                    .fill(Color.gray.opacity(0.3))
+                    .frame(width: 40, height: 40)
+
+                VStack(alignment: .leading, spacing: 4) {
+                    RoundedRectangle(cornerRadius: 2)
+                        .fill(Color.gray.opacity(0.3))
+                        .frame(height: 16)
+                        .frame(maxWidth: .infinity)
+
+                    RoundedRectangle(cornerRadius: 2)
+                        .fill(Color.gray.opacity(0.2))
+                        .frame(height: 12)
+                        .frame(maxWidth: .infinity * 0.7)
+                }
+
+                Spacer()
+            }
+            .padding(.vertical, 4)
+        }
+        .redacted(reason: .placeholder)
+    }
+}
