+++ Sources/DynamicList/Core Components/Dynamic List/DynamicListBuilder.swift
@@ -516,627 +516,6 @@ public final class DynamicListBuilder<Item: Identifiable & Hashable> {
     }
 }
 
-// MARK: - Sectioned DynamicList Builder
-
-/// A fluent builder class that simplifies the creation of SectionedDynamicList instances.
-///
-/// This builder provides a clean, chainable API for configuring and creating sectioned dynamic lists
-/// with support for static data, reactive publishers, custom UI components, and error handling.
-///
-/// ## Basic Usage
-/// ```swift
-/// SectionedDynamicListBuilder<Fruit>()
-///     .sections([
-///         ListSection(title: "Frutas Rojas", items: redFruits),
-///         ListSection(title: "Frutas Verdes", items: greenFruits)
-///     ])
-///     .rowContent { fruit in
-///         Text(fruit.name)
-///     }
-///     .detailContent { fruit in
-///         FruitDetailView(fruit: fruit)
-///     }
-///     .build()
-/// ```
-///
-/// ## Reactive Data Source
-/// ```swift
-/// SectionedDynamicListBuilder<Fruit>()
-///     .publisher(apiService.fetchFruitsByCategory())
-///     .title("Frutas por Categoría")
-///     .rowContent { fruit in
-///         FruitRowView(fruit: fruit)
-///     }
-///     .detailContent { fruit in
-///         FruitDetailView(fruit: fruit)
-///     }
-///     .build()
-/// ```
-///
-/// - Note: The `Item` type must conform to `Identifiable` and `Hashable` protocols.
-/// - Important: Use `build()` for standalone lists or `buildWithoutNavigation()` when embedding
-///   within existing navigation contexts to avoid NavigationStack conflicts.
-public final class SectionedDynamicListBuilder<Item: Identifiable & Hashable> {
-    // MARK: - Private Properties
-
-    /// Static sections to display in the list
-    private var sections: [ListSection<Item>] = []
-
-    /// Combine publisher for reactive data loading
-    private var publisher: AnyPublisher<[[Item]], Error>?
-
-    /// Custom row content builder
-    private var rowContent: ((Item) -> AnyView)?
-
-    /// Custom detail content builder
-    private var detailContent: ((Item) -> AnyView)?
-
-    /// Custom error content builder
-    private var errorContent: ((Error) -> AnyView)?
-
-    /// Custom skeleton content builder
-    private var skeletonContent: (() -> AnyView)?
-
-    /// Navigation title for the list
-    private var title: String?
-
-    /// Whether to hide the navigation bar
-    private var navigationBarHidden: Bool = false
-
-    // MARK: - Initialization
-
-    /// Creates a new SectionedDynamicListBuilder instance.
-    ///
-    /// Use this initializer to start building a sectioned dynamic list configuration.
-    /// All configuration is done through the fluent API methods.
-    public init() {}
-
-    // MARK: - Data Source Configuration
-
-    /// Sets static sections for the list.
-    ///
-    /// Use this method when you have a fixed array of sections that don't need to be loaded
-    /// from an external source. The sections will be displayed immediately without any loading state.
-    ///
-    /// - Parameter sections: The sections to display in the list.
-    /// - Returns: The builder instance for method chaining.
-    ///
-    /// ## Example
-    /// ```swift
-    /// SectionedDynamicListBuilder<Fruit>()
-    ///     .sections([
-    ///         ListSection(title: "Frutas Rojas", items: redFruits),
-    ///         ListSection(title: "Frutas Verdes", items: greenFruits)
-    ///     ])
-    ///     .rowContent { fruit in
-    ///         Text(fruit.name)
-    ///     }
-    ///     .build()
-    /// ```
-    @discardableResult
-    public func sections(_ sections: [ListSection<Item>]) -> Self {
-        self.sections = sections
-        return self
-    }
-
-    /// Sets static arrays for the list with optional titles.
-    ///
-    /// Use this method when you have arrays of arrays representing sections with optional titles.
-    ///
-    /// - Parameters:
-    ///   - arrays: Array of arrays representing sections
-    ///   - titles: Optional titles for each section
-    /// - Returns: The builder instance for method chaining.
-    ///
-    /// ## Example
-    /// ```swift
-    /// SectionedDynamicListBuilder<Fruit>()
-    ///     .arrays([redFruits, greenFruits, yellowFruits], titles: ["Rojas", "Verdes", "Amarillas"])
-    ///     .rowContent { fruit in
-    ///         Text(fruit.name)
-    ///     }
-    ///     .build()
-    /// ```
-    @discardableResult
-    public func arrays(_ arrays: [[Item]], titles: [String?] = []) -> Self {
-        let sections = zip(arrays, titles).map { items, title in
-            ListSection(title: title, items: items)
-        }
-        self.sections = sections
-        return self
-    }
-
-    /// Sets a Combine publisher for reactive data loading.
-    ///
-    /// Use this method when you want to load data from an external source like an API,
-    /// database, or any other reactive data stream. The publisher should emit arrays of arrays.
-    /// The list will automatically handle loading states, error handling, and data updates.
-    ///
-    /// - Parameter publisher: A Combine publisher that emits arrays of arrays.
-    /// - Returns: The builder instance for method chaining.
-    ///
-    /// ## Example
-    /// ```swift
-    /// SectionedDynamicListBuilder<Fruit>()
-    ///     .publisher(apiService.fetchFruitsByCategory())
-    ///     .rowContent { fruit in
-    ///         FruitRowView(fruit: fruit)
-    ///     }
-    ///     .build()
-    /// ```
-    @discardableResult
-    public func publisher(_ publisher: AnyPublisher<[[Item]], Error>) -> Self {
-        self.publisher = publisher
-        return self
-    }
-
-    // MARK: - Content Configuration
-
-    /// Sets custom row content for each item in the list.
-    ///
-    /// This method allows you to define how each item should be displayed in the list.
-    /// The content builder receives the item and should return a SwiftUI view.
-    ///
-    /// - Parameter content: A view builder that creates the row view for each item.
-    /// - Returns: The builder instance for method chaining.
-    @discardableResult
-    public func rowContent(@ViewBuilder _ content: @escaping (Item) -> some View) -> Self {
-        rowContent = { item in
-            AnyView(content(item))
-        }
-        return self
-    }
-
-    /// Sets custom detail content for when an item is selected.
-    ///
-    /// This method allows you to define the detail view that will be displayed when
-    /// a user taps on an item in the list. The content builder receives the selected item.
-    ///
-    /// - Parameter content: A view builder that creates the detail view for an item.
-    /// - Returns: The builder instance for method chaining.
-    @discardableResult
-    public func detailContent(@ViewBuilder _ content: @escaping (Item) -> some View) -> Self {
-        detailContent = { item in
-            AnyView(content(item))
-        }
-        return self
-    }
-
-    /// Sets custom error content for when loading fails.
-    ///
-    /// This method allows you to define a custom error view that will be displayed
-    /// when the data loading fails. The content builder receives the error that occurred.
-    ///
-    /// - Parameter content: A view builder that creates the error view.
-    /// - Returns: The builder instance for method chaining.
-    @discardableResult
-    public func errorContent(@ViewBuilder _ content: @escaping (Error) -> some View) -> Self {
-        errorContent = { error in
-            AnyView(content(error))
-        }
-        return self
-    }
-
-    /// Sets custom skeleton content for the loading state.
-    ///
-    /// Use this method to provide a custom skeleton view that will be displayed when the list
-    /// is in a loading state with no items available. This allows you to create skeleton
-    /// layouts that match your actual content structure.
-    ///
-    /// - Parameter content: A view builder that creates the skeleton view.
-    /// - Returns: The builder instance for method chaining.
-    @discardableResult
-    public func skeletonContent(@ViewBuilder _ content: @escaping () -> some View) -> Self {
-        skeletonContent = {
-            AnyView(content())
-        }
-        return self
-    }
-
-    /// Sets the navigation title for the list.
-    ///
-    /// - Parameter title: The title to display in the navigation bar.
-    /// - Returns: The builder instance for method chaining.
-    @discardableResult
-    public func title(_ title: String) -> Self {
-        self.title = title
-        return self
-    }
-
-    /// Hides the navigation bar.
-    ///
-    /// Use this method when you want to hide the navigation bar entirely.
-    /// Useful for full-screen experiences or when embedding in custom navigation.
-    ///
-    /// - Returns: The builder instance for method chaining.
-    @discardableResult
-    public func hideNavigationBar() -> Self {
-        navigationBarHidden = true
-        return self
-    }
-
-    // MARK: - Build Methods
-
-    /// Builds a complete SectionedDynamicList with NavigationStack.
-    ///
-    /// This method creates a fully functional sectioned dynamic list wrapped in its own NavigationStack.
-    /// Use this when you want a standalone list that manages its own navigation.
-    ///
-    /// - Returns: A SwiftUI view containing the configured sectioned dynamic list.
-    @MainActor
-    public func build() -> some View {
-        let viewModel: SectionedDynamicListViewModel<Item> = if let publisher {
-            SectionedDynamicListViewModel(dataProvider: { publisher }, initialSections: sections)
-        } else {
-            SectionedDynamicListViewModel(sections: sections)
-        }
-
-        return SectionedDynamicListWrapper(
-            viewModel: viewModel,
-            rowContent: rowContent ?? { item in
-                AnyView(DefaultRowView(item: item))
-            },
-            detailContent: detailContent ?? { item in
-                AnyView(DefaultDetailView(item: item))
-            },
-            errorContent: errorContent,
-            skeletonContent: skeletonContent,
-            title: title,
-            navigationBarHidden: navigationBarHidden,
-        )
-    }
-
-    /// Builds a SectionedDynamicList without NavigationStack for embedding in existing navigation.
-    ///
-    /// Use this method when you want to embed the list within an existing NavigationStack
-    /// or when you need to manage navigation at a higher level. This prevents navigation
-    /// conflicts and allows for more flexible navigation patterns.
-    ///
-    /// - Returns: A SwiftUI view containing the configured sectioned dynamic list without navigation wrapper.
-    @MainActor
-    public func buildWithoutNavigation() -> some View {
-        let viewModel: SectionedDynamicListViewModel<Item> = if let publisher {
-            SectionedDynamicListViewModel(dataProvider: { publisher }, initialSections: sections)
-        } else {
-            SectionedDynamicListViewModel(sections: sections)
-        }
-
-        return SectionedDynamicListContent(
-            viewModel: viewModel,
-            rowContent: rowContent ?? { item in
-                AnyView(DefaultRowView(item: item))
-            },
-            detailContent: detailContent ?? { item in
-                AnyView(DefaultDetailView(item: item))
-            },
-            errorContent: errorContent,
-            skeletonContent: skeletonContent,
-            title: title,
-            navigationBarHidden: navigationBarHidden,
-        )
-    }
-}
-
-// MARK: - DynamicList Wrapper
-
-/// Internal wrapper that provides NavigationStack for standalone lists.
-///
-/// This component is used internally by the `build()` method to wrap the list content
-/// in a NavigationStack. It's not meant to be used directly by consumers.
-private struct DynamicListWrapper<Item: Identifiable & Hashable>: View {
-    @State private var viewModel: DynamicListViewModel<Item>
-    private let rowContent: (Item) -> AnyView
-    private let detailContent: (Item) -> AnyView
-    private let errorContent: ((Error) -> AnyView)?
-    private let skeletonContent: (() -> AnyView)?
-    private let title: String?
-    private let navigationBarHidden: Bool
-
-    init(
-        viewModel: DynamicListViewModel<Item>,
-        rowContent: @escaping (Item) -> AnyView,
-        detailContent: @escaping (Item) -> AnyView,
-        errorContent: ((Error) -> AnyView)?,
-        skeletonContent: (() -> AnyView)?,
-        title: String?,
-        navigationBarHidden: Bool,
-    ) {
-        _viewModel = State(initialValue: viewModel)
-        self.rowContent = rowContent
-        self.detailContent = detailContent
-        self.errorContent = errorContent
-        self.skeletonContent = skeletonContent
-        self.title = title
-        self.navigationBarHidden = navigationBarHidden
-    }
-
-    var body: some View {
-        NavigationStack {
-            DynamicListContent(
-                viewModel: viewModel,
-                rowContent: rowContent,
-                detailContent: detailContent,
-                errorContent: errorContent,
-                skeletonContent: skeletonContent,
-                title: title,
-                navigationBarHidden: navigationBarHidden,
-            )
-        }
-    }
-}
-
-// MARK: - SectionedDynamicList Wrapper
-
-/// Internal wrapper that provides NavigationStack for standalone sectioned lists.
-///
-/// This component is used internally by the `build()` method to wrap the sectioned list content
-/// in a NavigationStack. It's not meant to be used directly by consumers.
-private struct SectionedDynamicListWrapper<Item: Identifiable & Hashable>: View {
-    @State private var viewModel: SectionedDynamicListViewModel<Item>
-    private let rowContent: (Item) -> AnyView
-    private let detailContent: (Item) -> AnyView
-    private let errorContent: ((Error) -> AnyView)?
-    private let skeletonContent: (() -> AnyView)?
-    private let title: String?
-    private let navigationBarHidden: Bool
-
-    init(
-        viewModel: SectionedDynamicListViewModel<Item>,
-        rowContent: @escaping (Item) -> AnyView,
-        detailContent: @escaping (Item) -> AnyView,
-        errorContent: ((Error) -> AnyView)?,
-        skeletonContent: (() -> AnyView)?,
-        title: String?,
-        navigationBarHidden: Bool,
-    ) {
-        _viewModel = State(initialValue: viewModel)
-        self.rowContent = rowContent
-        self.detailContent = detailContent
-        self.errorContent = errorContent
-        self.skeletonContent = skeletonContent
-        self.title = title
-        self.navigationBarHidden = navigationBarHidden
-    }
-
-    var body: some View {
-        NavigationStack {
-            SectionedDynamicListContent(
-                viewModel: viewModel,
-                rowContent: rowContent,
-                detailContent: detailContent,
-                errorContent: errorContent,
-                skeletonContent: skeletonContent,
-                title: title,
-                navigationBarHidden: navigationBarHidden,
-            )
-        }
-    }
-}
-
-// MARK: - DynamicList Content
-
-/// Internal component that renders the actual list content.
-///
-/// This component handles the core list rendering logic including loading states,
-/// error handling, and navigation. It's used by both `DynamicListWrapper` and
-/// the `buildWithoutNavigation()` method.
-private struct DynamicListContent<Item: Identifiable & Hashable>: View {
-    @State private var viewModel: DynamicListViewModel<Item>
-    private let rowContent: (Item) -> AnyView
-    private let detailContent: (Item) -> AnyView
-    private let errorContent: ((Error) -> AnyView)?
-    private let skeletonContent: (() -> AnyView)?
-    private let title: String?
-    private let navigationBarHidden: Bool
-
-    init(
-        viewModel: DynamicListViewModel<Item>,
-        rowContent: @escaping (Item) -> AnyView,
-        detailContent: @escaping (Item) -> AnyView,
-        errorContent: ((Error) -> AnyView)?,
-        skeletonContent: (() -> AnyView)?,
-        title: String?,
-        navigationBarHidden: Bool,
-    ) {
-        _viewModel = State(initialValue: viewModel)
-        self.rowContent = rowContent
-        self.detailContent = detailContent
-        self.errorContent = errorContent
-        self.skeletonContent = skeletonContent
-        self.title = title
-        self.navigationBarHidden = navigationBarHidden
-    }
-
-    var body: some View {
-        Group {
-            if viewModel.viewState.shouldShowLoading {
-                skeletonView
-            } else if viewModel.viewState.shouldShowError {
-                errorView
-            } else {
-                List(viewModel.viewState.items) { item in
-                    NavigationLink(value: item) {
-                        rowContent(item)
-                            .redacted(reason: viewModel.viewState.isLoading ? .placeholder : [])
-                    }
-                }
-                .refreshable {
-                    viewModel.refresh()
-                }
-            }
-        }
-        .navigationDestination(for: Item.self) { item in
-            detailContent(item)
-        }
-        .navigationTitle(title ?? "")
-        #if os(iOS)
-            .navigationBarHidden(navigationBarHidden)
-        #endif
-    }
-
-    @ViewBuilder
-    private var skeletonView: some View {
-        if let skeletonContent {
-            skeletonContent()
-        } else {
-            DefaultSkeletonView()
-        }
-    }
-
-    @ViewBuilder
-    private var errorView: some View {
-        if let errorContent,
-           let error = viewModel.viewState.error
-        {
-            errorContent(error)
-        } else if let error = viewModel.viewState.error {
-            DefaultErrorView(error: error)
-        }
-    }
-}
-
-// MARK: - SectionedDynamicList Content
-
-/// Internal component that renders the actual sectioned list content.
-///
-/// This component handles the core sectioned list rendering logic including loading states,
-/// error handling, and navigation. It's used by both `SectionedDynamicListWrapper` and
-/// the `buildWithoutNavigation()` method.
-private struct SectionedDynamicListContent<Item: Identifiable & Hashable>: View {
-    @State private var viewModel: SectionedDynamicListViewModel<Item>
-    private let rowContent: (Item) -> AnyView
-    private let detailContent: (Item) -> AnyView
-    private let errorContent: ((Error) -> AnyView)?
-    private let skeletonContent: (() -> AnyView)?
-    private let title: String?
-    private let navigationBarHidden: Bool
-
-    init(
-        viewModel: SectionedDynamicListViewModel<Item>,
-        rowContent: @escaping (Item) -> AnyView,
-        detailContent: @escaping (Item) -> AnyView,
-        errorContent: ((Error) -> AnyView)?,
-        skeletonContent: (() -> AnyView)?,
-        title: String?,
-        navigationBarHidden: Bool,
-    ) {
-        _viewModel = State(initialValue: viewModel)
-        self.rowContent = rowContent
-        self.detailContent = detailContent
-        self.errorContent = errorContent
-        self.skeletonContent = skeletonContent
-        self.title = title
-        self.navigationBarHidden = navigationBarHidden
-    }
-
-    var body: some View {
-        Group {
-            if viewModel.viewState.shouldShowLoading {
-                skeletonView
-            } else if viewModel.viewState.shouldShowError {
-                errorView
-            } else {
-                List {
-                    ForEach(viewModel.viewState.sections) { section in
-                        Section {
-                            ForEach(section.items) { item in
-                                NavigationLink(value: item) {
-                                    rowContent(item)
-                                        .redacted(reason: viewModel.viewState.isLoading ? .placeholder : [])
-                                }
-                            }
-                        } header: {
-                            if let title = section.title {
-                                Text(title)
-                            }
-                        } footer: {
-                            if let footer = section.footer {
-                                Text(footer)
-                            }
-                        }
-                    }
-                }
-                .refreshable {
-                    viewModel.refresh()
-                }
-            }
-        }
-        .navigationDestination(for: Item.self) { item in
-            detailContent(item)
-        }
-        .navigationTitle(title ?? "")
-        #if os(iOS)
-            .navigationBarHidden(navigationBarHidden)
-        #endif
-    }
-
-    @ViewBuilder
-    private var skeletonView: some View {
-        if let skeletonContent {
-            skeletonContent()
-        } else {
-            DefaultSectionedSkeletonView()
-        }
-    }
-
-    @ViewBuilder
-    private var errorView: some View {
-        if let errorContent,
-           let error = viewModel.viewState.error
-        {
-            errorContent(error)
-        } else if let error = viewModel.viewState.error {
-            DefaultErrorView(error: error)
-        }
-    }
-}
-
-// MARK: - Default Views
-
-/// Default row view used when no custom row content is provided.
-///
-/// This view displays a simple representation of the item with its description
-/// and ID. It's used as a fallback when `rowContent` is not specified.
-private struct DefaultRowView<Item: Identifiable & Hashable>: View {
-    let item: Item
-
-    var body: some View {
-        VStack(alignment: .leading) {
-            Text(String(describing: item))
-                .font(.body)
-            Text("ID: \(String(describing: item.id))")
-                .font(.caption)
-                .foregroundColor(.secondary)
-        }
-        .padding(.vertical, 4)
-    }
-}
-
-/// Default detail view used when no custom detail content is provided.
-///
-/// This view displays a simple detail representation of the item with its
-/// description and ID. It's used as a fallback when `detailContent` is not specified.
-private struct DefaultDetailView<Item: Identifiable & Hashable>: View {
-    let item: Item
-
-    var body: some View {
-        VStack(spacing: 16) {
-            Text(DynamicListPresenter.itemDetail)
-                .font(.largeTitle)
-                .fontWeight(.bold)
-
-            Text("\(String(describing: item))")
-                .font(.body)
-
-            Text("\(DynamicListPresenter.itemID): \(String(describing: item.id))")
-                .font(.caption)
-                .foregroundColor(.secondary)
-        }
-        .padding()
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-}
-
 // MARK: - Convenience Factory Methods
 
 /// Convenience factory methods for common use cases.
